refactor(frontend/user): create User into user module
