+++ packages/frontend/src/components/modules/User/User/UserProducts.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import '../../../styles/containers/userProducts.css';
-import { IProduct } from '../../../context/interfaces';
+import '../../../../styles/containers/userProducts.css';
+import { IProduct } from '../../../../context/interfaces';
 import ListProducts from './ListProducts';
 
 const UserProducts = ({ productsBought, cartProducts }: IUserProductsProps) => {
