+++ packages/frontend/src/components/molecules/index.ts
@@ -1,4 +1,4 @@
 import Searcher from './Searcher';
-import ListProducts from '../modules/User/ListProducts';
+import ListProducts from '../modules/User/User/ListProducts';
 
 export { Searcher, ListProducts };
